Updated build scripts for faster compiling during development

+++ build-scripts/build-leptonica.sh
@@ -27,8 +27,11 @@ CM_FLAGS=(
 echo "CM_FLAGS=${CM_FLAGS[@]}"
 
 cd $LIB_PATH
-rm -rf build
-mkdir -p build
+if [ $BUILD_CLEAN = 1 ]
+then
+  rm -rf build
+  mkdir -p build
+fi
 cd build
 emmake cmake .. -DCMAKE_C_FLAGS="$CXXFLAGS" ${CM_FLAGS[@]}
 emmake make install
